requirements.txt
beautifulsoup4==4.9.3
selenium==3.141.0
webdriver-manager==3.4.2
twilio
pyyaml==5.4.1
pyvirtualdisplay